+++ src/CFMS.Application/Features/FarmFeat/Update/UpdateFarmCommand.cs
@@ -6,7 +6,7 @@ namespace CFMS.Application.Features.FarmFeat.Update
 {
     public class UpdateFarmCommand : IRequest<BaseResponse<bool>>
     {
-        public UpdateFarmCommand(Guid farmId, string? farmName, string? farmCode, string? address, double? area, int? scale, string? phoneNumber, string? website, string? imageUrl)
+        public UpdateFarmCommand(Guid farmId, string? farmName, string? farmCode, string? address, decimal? area, int? scale, string? phoneNumber, string? website, string? imageUrl)
         {
             FarmId = farmId;
             FarmName = farmName;
@@ -27,7 +27,7 @@ namespace CFMS.Application.Features.FarmFeat.Update
 
         public string? Address { get; set; }
 
-        public double? Area { get; set; }
+        public decimal? Area { get; set; }
 
         public int? Scale { get; set; }
